Create LID_DRIVEN_CAVITY
As we know, the major difficulty encountered during solution of incompressible flow is the non-availability of the equation of state, which gives us a direct relationship between pressure and temperature. We need to find formulations that can couple pressure with the Navier Stokes equation. Pressure Poisson is one of the methods to solve the problem. There are many algorithms developed to couple Pressure Poisson equation with the Navier Stokes equations. One of them is SIMPLE (Semi Implicit Method for Pressure Correction) originally proposed by Patankar and Spalding [1] in 1972. A revised version of the SIMPLE proposed by Patankar, 1981 [3] which had no gradient of Pressure in the momentum equation. Similarly, SIMPLEC by Doormal and Kaithby,[3] are some of the methods developed to couple Pressure Poisson to Navier Stokes equation to solve for incompressible flow. Lid driven cavity is very widely referred model to test the numerical codes developed, as a lot of literature is available.  
The work presented here uses Pressure Poisson equation coupled to Navier Stokes equation to solve a laminar flow in square (LxL) driven cavity. SIMPLE method proposed by Patankar and Spalding 1972 used for pressure and velocity corrections. In addition, a scalr, fourth difference, third order accurate artificial dissipation for stability introduced. In the end solutions obtained compared with benchmarked solutions of Ghia et al [2]